#81 Adding the ability to generate code from a configuration file.
This is a work in progress - the 'generated' code from initialiseSafeToRun has highlighted an issue that should be addressed which is that it relies on building an object whih is in the 'internal' library. Instead, this should be re-written to generate the same code as exists in the register() functions instead of calling the register functions.